Remove Constants class, remove provider class, move classes to main package
