+++ src/main/java/com/prettysmarthomes/beaconscannerlib/BLeStartScanBroadcastReceiver.java
@@ -1,11 +1,8 @@
-package com.prettysmarthomes.beaconscannerlib.receiver;
+package com.prettysmarthomes.beaconscannerlib;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-
-import com.prettysmarthomes.beaconscannerlib.BLeScanService;
-import com.prettysmarthomes.beaconscannerlib.utils.Constants;
 
 
 public class BLeStartScanBroadcastReceiver extends BroadcastReceiver {
@@ -14,10 +11,10 @@ public class BLeStartScanBroadcastReceiver extends BroadcastReceiver {
     Intent queryIntent = new Intent(context, BLeScanService.class);
     queryIntent.putExtra(BLeScanService.EXTRA_SCAN_PERIOD,
         intent.getLongExtra(BLeScanService.EXTRA_SCAN_PERIOD,
-            Constants.DEFAULT_BLE_SCAN_PERIOD_MS));
+            ScanParameters.DEFAULT_BLE_SCAN_PERIOD_MS));
     queryIntent.putExtra(BLeScanService.EXTRA_SCAN_INTERVAL,
         intent.getLongExtra(BLeScanService.EXTRA_SCAN_INTERVAL,
-            Constants.DEFAULT_BLE_SCAN_INTERVAL_MS));
+            ScanParameters.DEFAULT_BLE_SCAN_INTERVAL_MS));
     queryIntent.putExtra(BLeScanService.EXTRA_FILTER_UUID,
         intent.getByteArrayExtra(BLeScanService.EXTRA_FILTER_UUID));
     context.startService(queryIntent);
