+++ src/main/java/com/prettysmarthomes/beaconscannerlib/di/BLeScanServiceComponent.java
@@ -1,0 +1,10 @@
+package com.prettysmarthomes.beaconscannerlib.di;
+
+import dagger.Component;
+
+/**
+ * Base component for dagger 2 for BLeScanService
+ */
+@Component(modules = BleScanServiceModule.class)
+public interface BLeScanServiceComponent {
+}
