--- src/main/java/com/prettysmarthomes/beaconscannerlib/provider/BluetoothAdapterProvider.java
@@ -1,11 +1,0 @@
-package com.prettysmarthomes.beaconscannerlib.provider;
-
-import android.bluetooth.BluetoothAdapter;
-
-
-public class BluetoothAdapterProvider {
-
-  public BluetoothAdapter getInstance() {
-    return BluetoothAdapter.getDefaultAdapter();
-  }
-}
