+++ src/test/java/com/prettysmarthomes/beaconscannerlib/CustomScanCallbackTest.java
@@ -4,6 +4,8 @@ import android.bluetooth.BluetoothDevice;
 import android.content.Context;
 import android.content.Intent;
 
+import com.prettysmarthomes.beaconscannerlib.view.CustomScanCallback;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
